Moves knownAbis to /src/helpers, now the /src/helpers/abi only contains .json files
